requirements.txt
click>=8.1.7
openai-agents>=0.0.17
python-dotenv>=1.0.0